Moved isweighted method from Edge class to Graph class.
As it is hard for me to imagine a graph can contain both weighted and
unweighted edges so far, I decided to not allow a graph to contain both.
Also, the logic which had checked if the given graph has both weighted and
unweighted edges or not was removed from Dijkstra class's
find_shortest_path method.